refactor(verify): split verify_a.rs at the SSM-dispatch seam to stay under the cap
#745's per-token verify metadata took verify_a.rs from 484 to 638 lines, and
`File size cap (<=500 LoC per .rs)` is a required context. Four other files in
this stack's diff are also over 500, but each was already over on main and is
allow-listed; this was the only new violation.

The four SSM state dispatch methods -- checkpoint, rollback, save_snapshot,
restore_snapshot -- move verbatim into verify_a_ssm.rs, leaving verify_a.rs at
417 and the sibling at 254. Same idiom as vision_encoder/enc_impl/: each sibling
adds methods to the same inherent impl, so dispatch and visibility are unchanged.

Allow-listing was the alternative and was rejected. The list's own rule requires
a rationale and a tracking issue, and its five entries are 505-528 -- marginally
over, awaiting a clean split. 638 is not marginal, and a clean seam existed.

The copy is verified exact, not asserted: extracting every method body from the
pre-split file and comparing against the two post-split files shows 6 methods in,
6 out, zero bodies differing. `cargo check --workspace --all-targets` EXIT=0.

Caught by the pipeline watcher on #870 rather than by me reading the diff -- the
LoC cap ran before I thought to.
